Courtyard football. It takes 6 people (3x3) to play football in the courtyard. The class Player contains the following fields: - “stamina” field, different for each instance - constants “MAX_STAMINA” with value 100 and “MIN_STAMINA” with value 0, common for all instances - static field countPlayers , which counts the number of players on the football field (initially they are 0, exit to the field is the creation of an instance of the class, leaving - when the player is tired).
and the following methods:
run() - The player runs when this method is called. This method reduces stamina by 1 when invoked once. When stamina reaches 0, the player needs to rest and leaves the field.
info() - displays a message depending on the number of players on the field. If there are less than 6 players, then it displays the message: “Teams are incomplete. There are still ... empty seats on the field”, else: “There are no empty seats on the field”.

Solution: I created the class Player according to the above pattern. An instance of a class, when created, has a hardiness value between 90 and 100 (generated inside the constructor). Created 6 players, called the method info(). When trying to create 7.8 ... n players, the number of players on the field does not change.
I applied the method run() to the player until he was completely exhausted (negative stamina is not allowed). Printed the number of players on the field to the console.